requirements.txt
accelerate 
aiohttp 
aiosignal
anyio 
archspec 
asttokens 
async-timeout 
attrs 
batchgenerators==0.25
boltons 
Bottleneck
Brotli 
cachetools 
certifi 
cffi 
charset-normalizer 
click 
colorama
comm 
conda-content-trust 
conda-package-handling 
conda_package_streaming 
contourpy 
coverage==6.5.0
coveralls==3.3.1
cryptography 
cycler 
dataclasses-json 
datasets 
debugpy
decorator 
dill 
distro
docopt==0.6.2
einops==0.7.0
et-xmlfile==1.1.0
exceptiongroup
executing 
faiss==1.7.4
filelock 
fonttools==4.25.0
frozenlist 
fsspec 
future==0.18.3
gmpy2 

greenlet 
huggingface-hub
idna
imagecodecs 
imageio 
importlib-metadata
importlib_resources==6.5.2
iniconfig 
ipykernel 
ipython 
jedi 
Jinja2 
joblib 
jsonpatch 
jsonpointer==2.1
jupyter_client 
jupyter_core 
kiwisolver 
kornia==0.4.1
lazy_loader 
libmambapy 
linecache2==1.0.0
MarkupSafe
marshmallow 
matplotlib 
matplotlib-inline 
menuinst 
mkl-fft 
mkl-random 
mkl-service==2.4.0
monai 
mpmath 
multidict 
multiprocess 
munkres==1.1.4
mypy-extensions 
nest-asyncio 
networkx 
nibabel==5.3.2
nltk 
numexpr 
numpy 
openai 
openapi-schema-pydantic 
opencv-python==4.8.1.78
openpyxl==3.0.10
packaging 
pandas 
pandas-stubs
parso 
pexpect
pickleshare
Pillow 
platformdirs 
plotly 
pluggy 
prompt-toolkit 
protobuf==3.20.3
psutil 
ptyprocess 
pure-eval 
pyarrow==11.0.0
pyasn1 
pyasn1-modules 
pycosat
pycparser 
pydantic 
pydantic_core 
pyparsing 
PySocks 
pytest 
pytest-mock==3.12.0
pytest-subtests 
python-dateutil 
pytz 
pyu2f 
PyWavelets 
PyYAML
pyzmq 
regex 
requests 
responses 
rsa 
ruamel.yaml 
safetensors 
scikit-image 
scikit-learn 
scipy 
sentence-transformers 
sentencepiece 
siren-pytorch==0.1.7
six 
sniffio 
SQLAlchemy 
stack-data 
sympy 
tenacity 
tensorboardX 
threadpoolctl 
tifffile 
timm 
tokenizers 
tomli 
torch==2.1.1
torch-lucent==0.1.8
torchaudio==2.1.1
torchvision==0.16.1
tornado 
tqdm 
traceback2==1.4.0
traitlets 
transformers 
triton==2.1.0
truststore 
types-pytz 
typing-inspect 
typing_extensions 
unittest2==1.1.0
urllib3
wcwidth 
xxhash 
yarl
zipp
zstandard 
